refactor: side effects parser plugin
